requirements.txt
cairocffi==0.8.0
CairoSVG==2.0.3
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
cssselect==1.0.1
cssselect2==0.2.0
Django==1.11.5
django-bootstrap3==9.0.0
django-crispy-forms==1.6.1
django-internationalflavor==0.3.1
django-widget-tweaks==1.4.1
html5lib==0.999999999
idna==2.6
lxml==4.0.0
olefile==0.44
Pillow==4.2.1
psycopg2==2.7.3.1
pycparser==2.18
Pyphen==0.9.4
pytz==2017.2
requests==2.18.4
six==1.11.0
tinycss==0.4
tinycss2==0.6.0
urllib3==1.22
WeasyPrint==0.40
webencodings==0.5.1
